requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.2

# LLM and AI
google-generativeai==0.3.2
langchain==0.1.6
langchain-google-genai==0.0.6
langchain-community==0.0.20
openai==1.10.0

# Database
duckdb==0.9.2
sqlalchemy==2.0.25

# Data visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Web search and external APIs
tavily-python==0.3.0
wikipedia==1.4.0
googletrans==3.1.0a0

# Geolocation
geopy==2.4.1
folium==0.15.1
streamlit-folium==0.16.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Text processing
tiktoken==0.5.1
sentence-transformers==2.3.1

# Data validation and formatting
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
loguru==0.7.2

# File handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Development tools
pytest==7.4.3
black==23.12.1
flake8==7.0.0
